frontend/requirements.txt
google-auth==2.35.0
requests>=2.32.0